Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I am not sure if we need this. I know the samsung hidl was introduced in 17.1.

indeed this is not needed. I just checked and stock does not have any thermal hal. the hidl impl just uses thermal.exynos5 if stock has it. what could be done, would be to use the thermal impl we added in hw/samsung in 17.1 but for that we need to figure out temperature values

Patch-set: 2
CC: Gerrit User 1000136 <1000136@c6d92919-bfe7-411c-9f64-004a8163e0f0>